add always_cache_features support
this feature is explained here: https://docs.puppetlabs.com/references/latest/configuration.html#alwayscachefeatures

default is false, in that case, puppet tries to load a feature on each run, even if the previous run failed to load it. you get a big performance boost if you enable this. than puppet checks it once and caches the result